requirements.txt
websocket-client==0.12.0
